requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-telegram-bot==21.5
aiofiles==24.1.0
slowapi==0.1.9
python-dotenv==1.0.1
pyyaml==6.0.2